# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BandExplanation": {
        "additionalProperties": false,
        "description": "Answers 'why this band?' for the queue tooltip and the case header.",
        "properties": {
          "band": {
            "$ref": "#/components/schemas/PriorityBand"
          },
          "basis": {
            "description": "How the band was reached, in working language.",
            "title": "Basis",
            "type": "string"
          },
          "caps_applied": {
            "default": [],
            "description": "Guards that limited the band, e.g. text similarity alone cannot reach the top band, or an incomplete bundle lowered certainty.",
            "items": {
              "type": "string"
            },
            "title": "Caps Applied",
            "type": "array"
          }
        },
        "required": [
          "band",
          "basis"
        ],
        "title": "BandExplanation",
        "type": "object"
      },
...
      "CaseDetailResponse": {
        "additionalProperties": false,
        "description": "Everything needed to understand and disposition one case.",
        "properties": {
          "band": {
            "$ref": "#/components/schemas/BandExplanation"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "case_version": {
            "minimum": 1.0,
            "title": "Case Version",
            "type": "integer"
          },
          "comparisons": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ComparisonCandidate"
            },
            "title": "Comparisons",
            "type": "array"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "encounter_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encounter End"
          },
          "encounter_start": {
            "format": "date-time",
            "title": "Encounter Start",
            "type": "string"
          },
          "evidence_completeness": {
            "$ref": "#/components/schemas/EvidenceCompleteness"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/ClaimLineView"
            },
            "title": "Lines",
            "type": "array"
          },
          "participant_token": {
            "title": "Participant Token",
            "type": "string"
          },
          "primary_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReasonDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "provider_token": {
            "title": "Provider Token",
            "type": "string"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/ReasonDto"
            },
            "title": "Reasons",
            "type": "array"
          },
          "sources": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SourceResource"
            },
            "title": "Sources",
            "type": "array"
          },
          "state": {
            "$ref": "#/components/schemas/CaseState"
          },
          "suggested_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What the system would suggest, e.g. REQUEST_EVIDENCE when the bundle is thin. A suggestion only \u2014 the reviewer chooses, and must still supply a reason.",
            "title": "Suggested Action"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            },
            "title": "Timeline",
            "type": "array"
          },
          "total_amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Amount",
            "type": "string"
          },
          "versions": {
            "$ref": "#/components/schemas/VersionStamp"
          }
        },
        "required": [
          "case_id",
          "case_version",
          "state",
          "participant_token",
          "provider_token",
          "total_amount",
          "currency",
          "encounter_start",
          "encounter_end",
          "primary_reason",
          "reasons",
          "band",
          "lines",
          "timeline",
          "evidence_completeness",
          "versions"
        ],
        "title": "CaseDetailResponse",
        "type": "object"
      },
      "CaseQueueResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "metrics": {
            "$ref": "#/components/schemas/QueueMetrics"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "metrics",
          "items",
          "page"
        ],
        "title": "CaseQueueResponse",
        "type": "object"
      },
      "CaseState": {
        "description": "Functional state model from the master plan \u00a7 10.",
        "enum": [
          "NEW",
          "SCREENED",
          "IN_REVIEW",
          "EVIDENCE_REQUESTED",
          "DISMISSED",
          "CONFIRMED_ANOMALY",
          "ESCALATED",
          "INVALID_INPUT"
        ],
        "title": "CaseState",
        "type": "string"
      },
      "CaseSummary": {
        "additionalProperties": false,
        "description": "One queue row.\n\n**Pseudonymous fields only, and no raw medical text** \u2014 that is a hard constraint from\n`docs/canonical/03_architecture.md`, asserted by `test_queue_response_carries_no_medical_text`.\nNarrative text appears only in the detail response, and then only the fragment a reason\ndepends on.\n\nField order mirrors the mandated column order: the sentence comes first.",
        "properties": {
          "band": {
            "$ref": "#/components/schemas/PriorityBand"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "case_version": {
            "minimum": 1.0,
            "title": "Case Version",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "evidence_completeness": {
            "$ref": "#/components/schemas/EvidenceCompleteness"
          },
          "modes": {
            "items": {
              "$ref": "#/components/schemas/RiskMode"
            },
            "title": "Modes",
            "type": "array"
          },
          "participant_token": {
            "description": "Pseudonymous. Never a real identifier.",
            "title": "Participant Token",
            "type": "string"
          },
          "provider_token": {
            "title": "Provider Token",
            "type": "string"
          },
          "reason_sentence": {
            "description": "Working language. The queue's first column.",
            "title": "Reason Sentence",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/CaseState"
          },
          "total_amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Amount",
            "type": "string"
          }
        },
        "required": [
          "reason_sentence",
          "modes",
          "case_id",
          "participant_token",
          "provider_token",
          "evidence_completeness",
          "total_amount",
          "currency",
          "created_at",
          "band",
          "state",
          "case_version"
        ],
        "title": "CaseSummary",
        "type": "object"
      },
      "ClaimLineView": {
        "additionalProperties": false,
        "description": "One billed line with its support state.\n\n`NOT_ASSESSABLE` is distinct from `UNSUPPORTED`: the first means the record was too thin\nto judge, the second means the evidence is genuinely absent. Conflating them is the\ndifference between requesting a document and alleging a service never happened.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "line_amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Line Amount",
            "type": "string"
          },
          "line_id": {
            "title": "Line Id",
            "type": "string"
          },
          "quantity": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity",
            "type": "string"
          },
          "service_at": {
            "format": "date-time",
            "title": "Service At",
            "type": "string"
          },
          "support_state": {
            "description": "SUPPORTED | PARTIALLY_SUPPORTED | UNSUPPORTED | NOT_ASSESSABLE",
            "title": "Support State",
            "type": "string"
          }
        },
        "required": [
          "line_id",
          "code",
          "description",
          "quantity",
          "line_amount",
          "service_at",
          "support_state"
        ],
        "title": "ClaimLineView",
        "type": "object"
      },
      "ComparisonCandidate": {
        "additionalProperties": false,
        "description": "Side-by-side pair for repeat-billing and cloned-documentation reasons.",
        "properties": {
          "candidate_case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Case Id"
          },
          "candidate_claim_id": {
            "title": "Candidate Claim Id",
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ComparisonField"
            },
            "title": "Fields",
            "type": "array"
          },
          "overlap_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlap End"
          },
          "overlap_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlap Start"
          },
          "similarity_components": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Similarity Components",
            "type": "object"
          },
          "template_caveat": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present on clone reasons. Legitimate template-based documentation produces high similarity too, and the reviewer must read that before acting.",
            "title": "Template Caveat"
          }
        },
        "required": [
          "candidate_claim_id",
          "fields"
        ],
        "title": "ComparisonCandidate",
        "type": "object"
      },
...
      "ErrorCode": {
        "description": "Stable identifiers. Append new members; never repurpose an existing one.",
        "enum": [
          "BUNDLE_TOO_LARGE",
          "BUNDLE_UNSUPPORTED_CONTENT_TYPE",
          "BUNDLE_DEPTH_EXCEEDED",
          "BUNDLE_MALFORMED_JSON",
          "BUNDLE_SCHEMA_INVALID",
          "BUNDLE_UNKNOWN_RESOURCE_TYPE",
          "BUNDLE_DANGLING_REFERENCE",
          "BUNDLE_CIRCULAR_REFERENCE",
          "BUNDLE_DUPLICATE_RESOURCE_ID",
          "BUNDLE_TOTAL_MISMATCH",
          "INGESTION_NOT_FOUND",
          "CASE_NOT_FOUND",
          "EVALUATION_RUN_NOT_FOUND",
          "BUNDLE_NOT_SCREENABLE",
          "DISPOSITION_REASON_REQUIRED",
          "CASE_VERSION_CONFLICT",
          "DISPOSITION_INVALID_TRANSITION",
          "AUDIT_FORBIDDEN",
          "CASE_ACCESS_FORBIDDEN",
          "CASE_REOPEN_FORBIDDEN",
          "USER_MANAGEMENT_FORBIDDEN",
          "SESSION_INVALID_CREDENTIALS",
          "SESSION_ACCOUNT_DEACTIVATED",
          "USER_NOT_FOUND",
          "USER_SELF_MODIFICATION_REFUSED",
          "USER_NO_CHANGE_REQUESTED",
          "BRIEFING_UNAVAILABLE"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "The single error envelope for every endpoint.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "issues": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Issues",
            "type": "array"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "EvidenceCompleteness": {
        "additionalProperties": false,
        "description": "How much of the expected supporting evidence was actually present.\n\nSurfaced in the queue so a reviewer can tell a thin record from a suspicious one before\nopening the case.",
        "properties": {
          "bundle_complete": {
            "title": "Bundle Complete",
            "type": "boolean"
          },
          "missing_reference_count": {
            "minimum": 0.0,
            "title": "Missing Reference Count",
            "type": "integer"
          },
          "supported_lines": {
            "minimum": 0.0,
            "title": "Supported Lines",
            "type": "integer"
          },
          "total_lines": {
            "minimum": 0.0,
            "title": "Total Lines",
            "type": "integer"
          }
        },
        "required": [
          "supported_lines",
          "total_lines",
          "missing_reference_count",
          "bundle_complete"
        ],
        "title": "EvidenceCompleteness",
        "type": "object"
      },
...
      "PageInfo": {
        "additionalProperties": false,
        "description": "Queue pagination. The full queue is never returned in one response.",
        "properties": {
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total_items": {
            "minimum": 0.0,
            "title": "Total Items",
            "type": "integer"
          },
          "total_pages": {
            "minimum": 0.0,
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total_items",
          "total_pages"
        ],
        "title": "PageInfo",
        "type": "object"
      },
      "PriorityBand": {
        "description": "Queue ordering bands.\n\n`NO_OBSERVED_RISK` means no selected detector fired at this engine version. It must never\nbe rendered as \"clean\" or \"safe\" \u2014 the system is not entitled to that claim.",
        "enum": [
          "DETERMINISTIC_CONFLICT",
          "HIGH_PRIORITY_SIGNAL",
          "NEEDS_CONTEXT",
          "NO_OBSERVED_RISK"
        ],
        "title": "PriorityBand",
        "type": "string"
      },
      "QueueMetrics": {
        "additionalProperties": false,
        "description": "The five operational metrics above the queue. Exactly five.\n\nAnything that does not change what a reviewer does next is excluded by\n`docs/canonical/01_product_decision.md` \u00a7 Main dashboard principles \u2014 no \"fraud saved\",\nno provider league tables, no national projections.",
        "properties": {
          "awaiting_review": {
            "minimum": 0.0,
            "title": "Awaiting Review",
            "type": "integer"
          },
          "deterministic_conflicts": {
            "minimum": 0.0,
            "title": "Deterministic Conflicts",
            "type": "integer"
          },
          "evidence_requested": {
            "minimum": 0.0,
            "title": "Evidence Requested",
            "type": "integer"
          },
          "median_time_in_queue_hours": {
            "minimum": 0.0,
            "title": "Median Time In Queue Hours",
            "type": "number"
          },
          "versions": {
            "$ref": "#/components/schemas/VersionStamp"
          }
        },
        "required": [
          "awaiting_review",
          "deterministic_conflicts",
          "evidence_requested",
          "median_time_in_queue_hours",
          "versions"
        ],
        "title": "QueueMetrics",
        "type": "object"
      },
      "ReasonCode": {
        "description": "Stable identifiers used by the UI and by tests. Never renumbered.",
        "enum": [
          "LINE_WITHOUT_COMPLETED_PROCEDURE",
          "LINE_WITHOUT_MEDICATION_DISPENSE",
          "SUPPORTING_EVIDENCE_ENTERED_IN_ERROR",
          "OVERLAPPING_CLAIM_SAME_EPISODE",
          "DUPLICATE_CLAIM_FINGERPRINT",
          "NEAR_DUPLICATE_DOCUMENTATION",
          "EPISODE_SPLIT_ACROSS_CLAIMS"
        ],
        "title": "ReasonCode",
        "type": "string"
      },
...
      "RiskMode": {
        "description": "The four officially listed facility risk modes this prototype addresses.",
        "enum": [
          "PHANTOM_OR_NO_PROCEDURE_EVIDENCE",
          "REPEAT_BILLING",
          "CLONED_DOCUMENTATION",
          "UNBUNDLING_FRAGMENTATION"
        ],
        "title": "RiskMode",
        "type": "string"
      },
...
      "SortKey": {
        "description": "What the reviewer chose to order the queue by.",
        "enum": [
          "band",
          "age",
          "amount",
          "evidence"
        ],
        "title": "SortKey",
        "type": "string"
      },
      "SortOrder": {
        "description": "Direction for the sortable queue columns. Ignored by the band sort \u2014 see `sort_cases`.",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder",
        "type": "string"
      },
...
      "SourceResource": {
        "additionalProperties": false,
        "description": "The resource behind an evidence reference, so every reference opens.\n\nAn unresolvable reference is recorded here with `MISSING` rather than dropped. Dropping it\nwould render as a shorter list \u2014 indistinguishable from a reason that cited less \u2014 which is\nexactly how a broken evidence trail hides.",
        "properties": {
          "availability": {
            "$ref": "#/components/schemas/SourceAvailability"
          },
          "fields": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SourceField"
            },
            "title": "Fields",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "resource_id": {
            "title": "Resource Id",
            "type": "string"
          },
          "resource_type": {
            "$ref": "#/components/schemas/ResourceType"
          }
        },
        "required": [
          "resource_type",
          "resource_id",
          "label",
          "availability"
        ],
        "title": "SourceResource",
        "type": "object"
      },
      "TimelineEvent": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "resource": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceRefDto"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "occurred_at",
          "kind",
          "label"
        ],
        "title": "TimelineEvent",
        "type": "object"
      },
...
      "ValidationIssue": {
        "additionalProperties": false,
        "description": "One problem found in a submitted bundle, pointed at a specific resource.\n\n`resource_type` and `resource_id` are what make the ingest screen's error list\nactionable \u2014 the operator learns *which* resource to fix, not merely that something is\nwrong somewhere.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "description": "Human-readable explanation. Never contains medical text.",
            "title": "Detail",
            "type": "string"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "resource_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Type"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "ValidationIssue",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Claim evidence integrity screening. Synthetic data only.",
    "title": "TilikKlaim API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/cases": {
      "get": {
        "description": "The ordered work list. Pseudonymous fields only; no narrative text.",
        "operationId": "list_cases_v1_cases_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CaseState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "band",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PriorityBand"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Band"
            }
          },
          {
            "in": "query",
            "name": "reason",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReasonCode"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RiskMode"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          },
          {
            "in": "query",
            "name": "created_after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created After"
            }
          },
          {
            "in": "query",
            "name": "created_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created Before"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortKey",
              "default": "band"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "desc"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "default": "reviewer",
              "title": "X-Actor-Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseQueueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Review queue",
        "tags": [
          "cases"
        ]
      }
    },
    "/v1/cases/{case_id}": {
      "get": {
        "description": "Claim lines, reasons with evidence and counter-evidence, timeline, and comparisons.",
        "operationId": "get_case_v1_cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "default": "reviewer",
              "title": "X-Actor-Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Case detail",
        "tags": [
          "cases"
        ]
      }
    },
...
  }
}